Edited README.md to include 'Hard' and 'Soft' dates for the relevant fields. Additionally deleted the meat information about the document.
Edited the lit-review to include basic information about four papers. Two of the papers include summary and relevant information pertaining to the works.

Amended the initial log to be used as a template.

Completed an initial log for the 1st week of using the reed-thesis Github repo. Included four main points for work in the upcoming week.